requirements.txt
flask
flask-restx
numpy
pandas
scikit-learn
geopy
geopandas
open3d-python
matplotlib